finished deleting all unused files
