+++ client/app.js
@@ -1,3 +1,5 @@
+
+
 import React from 'react'
 
 import {Navbar} from './components'
