--- client/components/main.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class MainPage extends Component {
-    render() {
-        return (
-            <div>
-                
-            </div>
-        );
-    }
-}
-
-export default MainPage;
