+++ client/index.js
@@ -4,8 +4,8 @@ import {Provider} from 'react-redux'
 import {Router} from 'react-router-dom'
 import history from './history'
 import store from './store'
-// import App from './app'
-import MainPage from './components/Root.jsx'
+import App from './app'
+
 
 // establishes socket connection
 import './socket'
@@ -13,7 +13,7 @@ import './socket'
 ReactDOM.render(
   <Provider store={store}>
     <Router history={history}>
-      <MainPage />
+      <App />
     </Router>
   </Provider>,
   document.getElementById('app')
