+++ client/store/index.js
@@ -4,12 +4,10 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 
-import products from './fetchProducts'
-import product from './fetchSingleProduct'
-import createProduct from './createProduct'
-import updateProduct from './updateProduct'
+import products from './productsReducer'
 
-const reducer = combineReducers({user, products, product, createProduct, updateProduct})
+
+const reducer = combineReducers({user, products})
 
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
@@ -19,7 +17,5 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
-export * from './fetchProducts'
-export * from './fetchSingleProduct'
-export * from './createProduct'
-export * from './updateProduct'
+export * from './productsReducer'
+
